+++ robuser/corruptions/__init__.py
@@ -1,0 +1,3 @@
+"""
+Module with various audio corruption methods
+"""
