+++ robuser/corruptions/corruption_type.py
@@ -1,12 +1,12 @@
-class NoiseGeneration:
+class CorruptionType:
     """
-    NoiseGeneration base class
+    CorruptionType base class
 
     """
 
     def __init__(self, config):
         """
-        Initialize the NoiseGeneration class
+        Initialize the CorruptionType class
 
         :param config: dictionary with the configuration parameters
         """
@@ -14,10 +14,10 @@ class NoiseGeneration:
 
     def run(self, audio_data, sample_rate):
         """
-        Run the augmentation method
+        Run the corruption method
 
         :param audio_data: numpy array with the audio data or the path to the audio file
         :param sample_rate: the sample rate of the audio data
-        :return: tuple with the augmented audio data (numpy array) and the applied noise (or None)
+        :return: tuple with the corrupted audio data (numpy array) and the applied noise (or None)
         """
         raise NotImplementedError
