+++ robuser/corruptions/gaussian.py
@@ -1,8 +1,8 @@
 from audiomentations import AddGaussianSNR
-from robuser.noise_generation.corruptions import NoiseGeneration
+from robuser.corruptions.corruption_type import CorruptionType
 
 
-class AWGNAugmentation(NoiseGeneration):
+class AWGNAugmentation(CorruptionType):
     """
     Class that augments the audio data with additive white Gaussian noise so that the signal-to-noise ratio (SNR) is
     equal to the specified value.
