+++ robuser/corruptions/get_corruption.py
@@ -1,9 +1,9 @@
-from robuser.noise_generation.content import ContentCorruption
-from robuser.noise_generation.gaussian import AWGNAugmentation
-from robuser.noise_generation.gain_transition import AddGainTransition
-from robuser.noise_generation.clipping_distortion import AddClippingDistortion
-from robuser.noise_generation.impulse_response import AddImpulseResponse
-from robuser.noise_generation.compression import Compression
+from robuser.corruptions.content import ContentCorruption
+from robuser.corruptions.gaussian import AWGNAugmentation
+from robuser.corruptions.gain_transition import AddGainTransition
+from robuser.corruptions.clipping_distortion import AddClippingDistortion
+from robuser.corruptions.impulse_response import AddImpulseResponse
+from robuser.corruptions.compression import Compression
 
 
 def get_corruption(corruption_name):
