--- robuser/noise_generation/__init__.py
@@ -1,3 +1,0 @@
-"""
-Module with various noise generation methods
-"""
